Relocate icon and nav bar files into subfolders for maintenance purposes
