+++ app/client/main.tsx
@@ -6,9 +6,9 @@ import { AboutPage } from "/ui-modules/about/AboutPage";
 import { BrowserRouter, Routes, Route } from "react-router";
 import { Provider } from "react-redux";
 import { store } from "/app/store";
-import { TopNavbar } from "/ui-modules/shared/Navbar";
-import { BottomNavbar } from "/ui-modules/shared/BottomNavbar";
-import { SideNavBar } from "/ui-modules/shared/SideNavbar";
+import { TopNavbar } from "../../ui-modules/shared/navigation-bars/Navbar";
+import { BottomNavbar } from "../../ui-modules/shared/navigation-bars/BottomNavbar";
+import { SideNavBar } from "../../ui-modules/shared/navigation-bars/SideNavbar";
 
 Meteor.startup(initialiseReactRoot);
 
@@ -31,7 +31,7 @@ function AppRoot(): React.JSX.Element {
         <BrowserRouter>
           {/* TopNavbar: Pass `setSidebarOpen` prop to control sidebar state */}
           <TopNavbar setSidebarOpen={setSidebarOpen} />
-          
+
           <Routes>
             <Route path="/" element={<ExampleHomePage />} />
             <Route path="/about" element={<AboutPage />} />
@@ -39,7 +39,10 @@ function AppRoot(): React.JSX.Element {
           </Routes>
 
           {/* Sidebar & Backdrop */}
-          <SideNavBar isOpen={isSidebarOpen} onClose={() => setSidebarOpen(false)} />
+          <SideNavBar
+            isOpen={isSidebarOpen}
+            onClose={() => setSidebarOpen(false)}
+          />
           <BottomNavbar />
         </BrowserRouter>
       </Provider>
