+++ ui-modules/shared/icons/EscapeIcon.tsx
