+++ ui-modules/shared/icons/MultipleHousesIcon.tsx
