+++ ui-modules/shared/icons/PropManagerLogoIcon.tsx
